drivers: pwm: add PWM pattern to header
Add the possibility to send patterns to the PWM drivers. This could
allow for the drivers to generate a pattern of PWM signals to be send
out one after another without interference from the system. It could
be useful for protocols that are not implemented in hardware like the
WS2812b LED strips or DALI telegrams.

Signed-off-by: Fabian Pflug <fabian.pflug@gumulka.eu>